autogen-copilot/requirements.txt
ipython
azure-search-documents 
Flask
flask[async]
Flask-Cors
ipykernel
python-dotenv
pandas
pyodbc
sqlalchemy
tenacity
pyautogen~=0.2.0b4
aiohttp
azure-ai-formrecognizer
